python/requirements.txt
luigi
grpcio==1.46.3 
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
python-dotenv